Base Renderer Util
