+++ lib/br_invoices_pdf.rb
@@ -47,5 +47,7 @@ require 'br_invoices_pdf/util/xml_locate'
 require 'br_invoices_pdf/util/nfce_check_urls'
 require 'br_invoices_pdf/util/mount_params'
 require 'br_invoices_pdf/util/enum'
+require 'br_invoices_pdf/util/base_renderer'
+require 'br_invoices_pdf/util/pdf_renderer'
 require 'br_invoices_pdf/cfe'
 require 'br_invoices_pdf/nfce'
