+++ lib/br_invoices_pdf/cfe/renderer.rb
@@ -14,7 +14,7 @@ module BrInvoicesPdf
     module Renderer
       module_function
 
-      AUTO_HEIGHT_MOCK = 2000
+      extend Util::PdfRenderer
 
       RENDERERS = [
         CompanyIdentification,
@@ -27,24 +27,8 @@ module BrInvoicesPdf
         FiscoInfo
       ].freeze
 
-      # :reek:FeatureEnvy
       def pdf(data, options)
-        page_width = Renderer::BaseRenderer.page_paper_width(options[:page_size])
-
-        Prawn::Document.new(options.merge(page_size: [page_width, AUTO_HEIGHT_MOCK])) do |pdf|
-          pdf_content(pdf, data, page_width)
-        end
-      end
-
-      def pdf_content(pdf, data, page_width)
-        pdf.font_size(7) do
-          RENDERERS.each do |renderer|
-            renderer.execute(pdf, data)
-          end
-
-          page = pdf.page
-          page.dictionary.data[:MediaBox] = [0, pdf.y - page.margins[:bottom], page_width, AUTO_HEIGHT_MOCK]
-        end
+        generate_pdf(data, options, RENDERERS)
       end
     end
   end
