+++ lib/br_invoices_pdf/cfe/renderer/company_identification.rb
@@ -5,6 +5,7 @@ module BrInvoicesPdf
     module Renderer
       # :reek:DataClump
       module CompanyIdentification
+        extend Util::BaseRenderer
         extend BaseRenderer
 
         module_function
