+++ lib/br_invoices_pdf/cfe/renderer/fisco_info.rb
@@ -4,6 +4,7 @@ module BrInvoicesPdf
   module Cfe
     module Renderer
       module FiscoInfo
+        extend Util::BaseRenderer
         extend BaseRenderer
 
         module_function
