+++ lib/br_invoices_pdf/cfe/renderer/header.rb
@@ -4,6 +4,7 @@ module BrInvoicesPdf
   module Cfe
     module Renderer
       module Header
+        extend Util::BaseRenderer
         extend BaseRenderer
 
         module_function
