+++ lib/br_invoices_pdf/cfe/renderer/payment_forms.rb
@@ -4,6 +4,7 @@ module BrInvoicesPdf
   module Cfe
     module Renderer
       module PaymentForms
+        extend Util::BaseRenderer
         extend BaseRenderer
 
         module_function
