+++ lib/br_invoices_pdf/cfe/renderer/product_table.rb
@@ -4,6 +4,7 @@ module BrInvoicesPdf
   module Cfe
     module Renderer
       module ProductTable
+        extend Util::BaseRenderer
         extend BaseRenderer
 
         module_function
