+++ lib/br_invoices_pdf/cfe/renderer/taxes_info.rb
@@ -4,6 +4,7 @@ module BrInvoicesPdf
   module Cfe
     module Renderer
       module TaxesInfo
+        extend Util::BaseRenderer
         extend BaseRenderer
 
         module_function
