+++ lib/br_invoices_pdf/nfce/renderer.rb
@@ -13,9 +13,9 @@ require 'br_invoices_pdf/nfce/renderer/qr_code'
 module BrInvoicesPdf
   module Nfce
     module Renderer
-      module_function
+      extend Util::PdfRenderer
 
-      AUTO_HEIGHT_MOCK = 2000
+      module_function
 
       RENDERERS = [
         CompanyIdentification,
@@ -29,24 +29,8 @@ module BrInvoicesPdf
         QrCode
       ].freeze
 
-      # :reek:FeatureEnvy
       def pdf(data, options)
-        page_width = Renderer::BaseRenderer.page_paper_width(options[:page_size])
-
-        Prawn::Document.new(options.merge(page_size: [page_width, AUTO_HEIGHT_MOCK])) do |pdf|
-          pdf_content(pdf, data, page_width)
-        end
-      end
-
-      def pdf_content(pdf, data, page_width)
-        pdf.font_size(7) do
-          RENDERERS.each do |renderer|
-            renderer.execute(pdf, data)
-          end
-
-          page = pdf.page
-          page.dictionary.data[:MediaBox] = [0, pdf.y - page.margins[:bottom], page_width, AUTO_HEIGHT_MOCK]
-        end
+        generate_pdf(data, options, RENDERERS)
       end
     end
   end
