+++ lib/br_invoices_pdf/nfce/renderer/customer_identification.rb
@@ -4,6 +4,7 @@ module BrInvoicesPdf
   module Nfce
     module Renderer
       module CustomerIdentification
+        extend Util::BaseRenderer
         extend BaseRenderer
 
         module_function
