+++ lib/br_invoices_pdf/nfce/renderer/fiscal_message.rb
@@ -4,6 +4,7 @@ module BrInvoicesPdf
   module Nfce
     module Renderer
       module FiscalMessage
+        extend Util::BaseRenderer
         extend BaseRenderer
 
         module_function
