+++ lib/br_invoices_pdf/nfce/renderer/product_table.rb
@@ -4,6 +4,7 @@ module BrInvoicesPdf
   module Nfce
     module Renderer
       module ProductTable
+        extend Util::BaseRenderer
         extend BaseRenderer
 
         module_function
