+++ lib/br_invoices_pdf/nfce/renderer/taxes_info.rb
@@ -4,6 +4,7 @@ module BrInvoicesPdf
   module Nfce
     module Renderer
       module TaxesInfo
+        extend Util::BaseRenderer
         extend BaseRenderer
 
         module_function
