+++ lib/br_invoices_pdf/nfce/renderer/totals.rb
@@ -4,6 +4,7 @@ module BrInvoicesPdf
   module Nfce
     module Renderer
       module Totals
+        extend Util::BaseRenderer
         extend BaseRenderer
 
         module_function
